==> 1_classification/1_exploring.ipynb <==
# Exploring the Titanic dataset

We'll start off by taking a look at our data set.

The first step in any notebook (or python script) is to import any dependancies. 

In this notebook we'll be using the pandas library to read our CSV file, hold it as a dataframe within our notebook so that we can easily manipulate the data.

A dataframe is a 2-dimensional labeled data structure with columns and rows. Think of it like a spreadsheet to hold our data within the python code.

We import the pandas libray with the `import` keyword, we then set an alias for it by using the `as` keyword. 

This is not required but it is a convention that is widely used, this allows us to access the pandas library functions 

using `pd` instead of typing out `pandas` each time.

import pandas as pd 

### Dataset description

Here we have a description of the column headers of the CSV data.

1. PassengerId - Passenger unique Id
2.  Survival - Survival (0 = No; 1 = Yes).
3.  Pclass - Passenger Class (1 = 1st; 2 = 2nd; 3 = 3rd)
4.  Name - Name
5.  Sex - Sex
6.  Age - Age
7.  Sibsp - Number of Siblings/Spouses Aboard
8.  Parch - Number of Parents/Children Aboard
9.  Ticket - Ticket Number
10.  Fare - Passenger Fare
11. Cabin - Cabin
12. Embarked - Port of Embarkation (C = Cherbourg; Q = Queenstown; S = Southampton)

We use the `.read_csv()` function from the pandas library. This function require 1 argument, the path to a csv file.

We declare a variable `titanic_df` that our data will be saved to, we can then reference this variable when we want to perform operations on our data.

titanic_df = pd.read_csv('datasets/titanic.csv')

We next use the `.head()` method to print out the first n rows of the dataframe.

We can supply an argument of an integer number to print out that number of rows.

titanic_df.head(10)

Try to show the top 10 rows of the dataset.

# Write your code here

We can look at the `shape` attribute of our dataframe to see the number of rows and columns.

titanic_df.shape

We have 891 records and 12 columns

We can also call the `.count()` function, this will tell us how many non-null values are in each column.

titanic_df.count()

As we can see there are a few columns with missing data, we will look into how we clean this up in the next notebook.

### Get Some Statistics

To view some basic statistical details like percentile, mean, std etc. of the numeric columns present in the dataframe,We have used
.describe() function of pandas library

titanic_df.describe()

### Get the count of number of survivors

As motive of this study is that, given a new passenger detail we can predict if a passenger will survive or not and we have a binary variable 'survived' which is providing info for survival. Let's see how many of the passenger available in the dataset actually survived on the titanic.

.value_counts() function returns object containing counts of unique values, excluding NA values by default.

titanic_df['Survived'].value_counts()

### Visualize the count of survivors

For visualizing the survivors, we have used Seaborn library.It is a Python data visualization library based on matplotlib. It provides a high-level interface for drawing attractive and informative statistical graphics.

As we did import pd in the start, in the same way we can import seaborn using 'import' keyword.

import seaborn as sns

sns.countplot(titanic_df['Survived'])

Similiary using above code we can do countplot for variable ('Parch','Embarked',Pclass,'Sex',Sibsp).
